split rust code to binary and library and refactor to use enum
