+++ Cargo.toml
@@ -1,17 +1,21 @@
 [package]
 name = "skalsa_bot"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-anyhow = "1.0.82"
+anyhow = "1.0.83"
 chrono = { version = "0.4.38", features = ["serde"] }
 clap = { version = "4.5.4", features = ["derive"] }
 colored = "2.1.0"
-reqwest = { version = "0.12.3", features = ["json"] }
-serde = { version = "1.0.198", features = ["derive"] }
-serde_json = "1.0.116"
-strum = { version = "0.26", features = ["derive"] }
+reqwest = { version = "0.12.4", features = ["json"] }
+serde = { version = "1.0.202", features = ["derive"] }
+serde_json = "1.0.117"
+strum = { version = "0.26.2", features = ["derive"] }
 tokio = { version = "1.37.0", features = ["rt", "rt-multi-thread", "macros"] }
+
+[[bin]]
+name = "skalsa"
+path = "src/main.rs"
